pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langgraph-cb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "gradio>=5.0.0",
    "langchain>=1.2.7",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "langsmith>=0.6.4",
    "notebook>=7.5.2",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[project.scripts]
langgraph-cb = "langgraph_cb.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
